refactor: organize aplication enviroment and project structure
