--- src/controller/index.js
@@ -1,6 +1,0 @@
-import getReceitaCnpj from "./receita.controller";
-import getSituacaoCnpj from "./situacao.controller";
-import getConsultarCnpj from "./consultar.controller";
-import getCnpj from "./cnpj.controller";
-
-export { getReceitaCnpj, getSituacaoCnpj, getConsultarCnpj, getCnpj };
