+++ src/controllers/index.js
@@ -1,0 +1,11 @@
+import ReceitaController from "./receita.controller";
+import SituacaoController from "./situacao.controller";
+import ConsultarController from "./consultar.controller";
+import CnpjController from "./cnpj.controller";
+
+export {
+  ReceitaController,
+  SituacaoController,
+  ConsultarController,
+  CnpjController,
+};
