+++ src/services/index.js
@@ -1,0 +1,6 @@
+import ReceitaService from "./receita.services";
+import ConsultarService from "./consultar.services";
+import SituacaoServices from "./situacao.services";
+import CnpjService from "./cnpj.services";
+
+export { ReceitaService, ConsultarService, SituacaoServices, CnpjService };
